Fix presentation attributes in fieldbuilder
