+++ api/src/test/scala/com/tersesystems/echopraxia/plusscala/api/Logging.scala
@@ -1,4 +1,8 @@
 package com.tersesystems.echopraxia.plusscala.api
 
+import java.time.Instant
+
 // This trait should be extended for domain model classes
-trait Logging extends LoggingBase
+trait Logging extends LoggingBase {
+  implicit val instantToValue: ToValue[Instant] = instant => ToValue(instant.toString)
+}
